% Generated by roxygen2 and by hand
\name{xrf_entry}
\alias{xrf_entry}
\title{Graphical user interface for XRF data entry}
\usage{
xrf_entry()
}
\arguments{
\item{xrf_entry}{Shine app for XRF databse}
}
\value{
}
\description{
This package provide a graphical user interface to enter XRF data easily into the database.
}
\examples{
xrf_entry()
}
\seealso{
}
